Revert "feat: 버저닝 추가"
This reverts commit a45d3f15fce4be8de5dbd8da8a7511f1803fa6f0.
